Clean up unneeded function, move query to volumes.ts
