+++ packages/manager/src/utilities/pricing/dynamicPricing.test.ts
@@ -9,9 +9,6 @@ import {
   getDCSpecificPriceByType,
   renderMonthlyPriceToCorrectDecimalPlace,
 } from './dynamicPricing';
-import { getDynamicVolumePrice } from './dynamicVolumePrice';
-
-const mockVolumeType = volumeTypeFactory.build();
 
 describe('getDCSpecificPricingDisplay', () => {
   it('calculates dynamic pricing for a region without an increase', () => {
@@ -39,16 +36,6 @@ describe('getDCSpecificPricingDisplay', () => {
     ).toBe('28.00');
   });
 
-  it('calculates dynamic pricing for a volumes based on size', () => {
-    expect(
-      getDynamicVolumePrice({
-        regionId: 'id-cgk',
-        size: 20,
-        type: mockVolumeType,
-      })
-    ).toBe('2.40');
-  });
-
   it('handles default case correctly', () => {
     expect(
       getDCSpecificPrice({
@@ -61,6 +48,7 @@ describe('getDCSpecificPricingDisplay', () => {
 
 describe('getDCSpecificPricingByType', () => {
   const mockNodeBalancerType = nodeBalancerTypeFactory.build();
+  const mockVolumeType = volumeTypeFactory.build();
 
   it('calculates dynamic pricing for a region without an increase', () => {
     expect(
@@ -85,6 +73,16 @@ describe('getDCSpecificPricingByType', () => {
         type: mockNodeBalancerType,
       })
     ).toBe('14.00');
+  });
+
+  it('calculates dynamic pricing for a volume based on size', () => {
+    expect(
+      getDCSpecificPriceByType({
+        regionId: 'id-cgk',
+        size: 20,
+        type: mockVolumeType,
+      })
+    ).toBe('2.40');
   });
 
   it('handles an invalid price if region is not available', () => {
